Combined PR of some QoL features and small code refactors (#527)
I combined some smaller QoL features and code refactors into one PR to
ease and speed up the reviewing proces.
I think I made this branch off the `move models to models folder` branch
before it was merged, so that is why those commits are there.
This PR does the following:
- Refactor some model relationships to use the `Model::class` notation
instead of the `'path/to/model'` notation #521
- Adds some typehints for (repository) variables in controllers
- Add a lot of confirmation dialogs for deleting resources #467
- Only show the "*mandatory question" span in the application form if
there are actually mandatory questions #524
- Add the subscription end date to the agenda item details view #373

---------

+++ resources/views/beheer/mailList/show.blade.php
@@ -18,7 +18,10 @@
                 <a href="{{url('/mailList/')}}" class="btn btn-primary">
                     <em class="ion-android-arrow-back"></em>  {{'Back'}}
                 </a>
-                <form method="post" action="{{ url('mailList/' . $mailList->getId()) }}" onsubmit="return confirm('Are you sure you want to delete the maillist?');">
+                <form method="post"
+                      action="{{ url('mailList/' . $mailList->getId()) }}"
+                      onsubmit="return confirm('Are you sure you want to delete the maillist?');"
+                >
                     @csrf
                     @method('DELETE')
                     <button type="submit" class="btn btn-danger btn-primary"><em class="ion-trash-a"></em> {{'Remove'}}</button>
@@ -70,7 +73,13 @@
                         <td>{{$member->getAddress()}}</td>
                         <td>{{$member->getName()}}</td>
                         <td>
-                            <a href="#" id="deleteMember" data-mailList-id="{{$mailList->getId()}}" data-member-email="{{$member->getAddress()}}"><em class="ion-trash-a"></em></a>
+                            <a href="#"
+                               id="deleteMember"
+                               data-mailList-id="{{$mailList->getId()}}"
+                               data-member-email="{{$member->getAddress()}}"
+                            >
+                                <em class="ion-trash-a"></em>
+                            </a>
                         </td>
                     </tr>
                 @endforeach
@@ -100,17 +109,20 @@
         $(document).on('click', '#deleteMember', function() {
             const mailListId = $(this).attr('data-mailList-id');
             const memberEmail = $(this).attr('data-member-email');
-
-            $.ajax({
-                url: '/mailList/' + mailListId + '/member/' + memberEmail,
-                data: {
-                    _token: window.Laravel.csrfToken
-                },
-                type: 'DELETE',
-                success: function() {
-                    window.location.reload();
-                }
-            });
+            
+            if (confirm('Are you sure you want to remove this member from {{$mailList->getAddress()}}?'))
+            {
+                $.ajax({
+                    url: '/mailList/' + mailListId + '/member/' + memberEmail,
+                    data: {
+                        _token: window.Laravel.csrfToken
+                    },
+                    type: 'DELETE',
+                    success: function() {
+                        window.location.reload();
+                    }
+                });
+            }
         });
     </script>
 @endpush
